Move `_TrainingEpochLoop`
